Separate logging and reqid options (#1164)
* Separate logging and reqid options #1035

* Updated based on comments
