+++ docs/Logging.md
@@ -62,17 +62,7 @@ fastify.get('/', options, function (request, reply) {
 
 <a name="logging-request-id" />
 
-By default fastify adds an id to every request for easier tracking. If the "request-id" header is present its value is used, otherwise a new incremental id is generated. See Fastify Factory [`requestIdHeader`](https://github.com/fastify/fastify/blob/master/docs/Server.md#factory-request-id-header) options for customizing that header name.
-Additionally, `genReqId` option can be used for generating the request id by yourself. It will received the incoming request as a parameter.
-
-```js
-let i = 0
-const fastify = require('fastify')({
-  logger: {
-    genReqId: function (req) { return i++ }
-  }
-})
-```
+By default fastify adds an id to every request for easier tracking. If the "request-id" header is present its value is used, otherwise a new incremental id is generated. See Fastify Factory [`requestIdHeader`](https://github.com/fastify/fastify/blob/master/docs/Server.md#factory-request-id-header) and Fastify Factory [`genReqId`](https://github.com/fastify/fastify/blob/master/docs/Server.md#gen-request-id) for customization options.
 
 The default logger is configured with a set of standard serializers that serialize objects with `req`, `res`, and `err` properties. This behavior can be customized by specifying custom serializers.
 ```js
