+++ docs/Server.md
@@ -154,6 +154,21 @@ The header name used to know the request id. See [the request id](https://github
 
 + Default: `'request-id'`
 
+<a name="factory-gen-request-id"></a>
+### `genReqId`
+
+Function for generating the request id. It will receive the incoming request as a parameter.
+
++ Default: `value of 'request-id' if provided or monotonically increasing integers`
+
+Especially in distributed systems, you may want to override the default id generation behaviour as shown below. For generating `UUID`s you may want to checkout [hyperid](https://github.com/mcollina/hyperid)
+
+```js
+let i = 0
+const fastify = require('fastify')({
+  genReqId: function (req) { return i++ }
+})
+
 <a name="factory-trust-proxy"></a>
 ### `trustProxy`
