+++ lib/logger.js
@@ -43,14 +43,6 @@ function createLogger (opts, stream) {
   return logger
 }
 
-function reqIdGenFactory (requestIdHeader) {
-  var maxInt = 2147483647
-  var nextReqId = 0
-  return function genReqId (req) {
-    return req.headers[requestIdHeader] || (nextReqId = (nextReqId + 1) & maxInt)
-  }
-}
-
 const serializers = {
   req: function asReqValue (req) {
     return {
@@ -77,7 +69,6 @@ function now () {
 
 module.exports = {
   createLogger,
-  reqIdGenFactory,
   serializers,
   now
 }
